added coments and likes

+++ src/components/DeleteBlog.jsx
@@ -6,18 +6,24 @@ import { deleteObject, ref } from 'firebase/storage';
 
 export default function DeleteBlog({id, imageUrl}) {
     const handleDelete = async() => {
-        try {
-            await deleteDoc(doc(db, "Articles", id))
-            toast("Blog deleted successfully", {type: "success"})
-            const storageRef = ref(storage, imageUrl);
-            await deleteObject(storageRef)
-        } catch (error) {
-            toast("Error deleting blog", {type: "error"})
-        }
-    }
-  return (
+        if (window.confirm("Are you sure you want to delete this article?")) {
+                try {
+                    await deleteDoc(doc(db, "Articles", id))
+                    toast("Blog deleted successfully", {type: "success"})
+                    const storageRef = ref(storage, imageUrl);
+                    await deleteObject(storageRef)
+                } catch (error) {
+                    toast("Error deleting blog", {type: "error"})
+                }
+            }
+    };
+    return (
     <div>
-        <button className='btn btn-danger' onClick={handleDelete}>Delete</button>
+      <i
+        className="fa fa-times"
+        onClick={handleDelete}
+        style={{ cursor: "pointer" }}
+      /> 
     </div>
-  )
+    )
 }
